connect App to redux store (disconnect PhotoGallery)
